+++ codexcost/watcher.py
@@ -23,13 +23,13 @@ def watch(start_timestamp:datetime, output:Callable[[Iterable[TokenCount]],None]
     try:
         # seed with current sessions to build initial sessions state index
         for session in _watch_session_changes(base_path, seed=find_sessions(base_path)):
-            size, state = sessions.get(session, (0, SessionContext()))
+            size, context = sessions.get(session, (0, SessionContext(session)))
             new_size = session.stat().st_size
 
             # modifications flushed changed to file so parse new tokens
             if new_size > size:
                 try:
-                    gen = parse_session(session, session_state=state)
+                    gen = parse_session(context)
                     
                     while True:
                         count = next(gen)
@@ -37,10 +37,10 @@ def watch(start_timestamp:datetime, output:Callable[[Iterable[TokenCount]],None]
                         if count.timestamp >= start_timestamp:
                             output([count])
                 except StopIteration as e:
-                    state = e.value
+                    context = e.value
                 
                 # store changed session infos
-                sessions[session] = (new_size, state)
+                sessions[session] = (new_size, context)
     except KeyboardInterrupt:
         pass